Updated tearcher.cs , studentInfo.cs.
In teacher.cs file check the null value before updating value in updateStudent and DeleteStudent method.
Renamed ManageUSerThings to ManagePerson and in that check for option value.
In StudentIinfo class check the null value in of student.
