+++ Practical8/Practical8/Practical8.csproj
@@ -52,7 +52,7 @@
     <Compile Include="IStudentInfo.cs" />
     <Compile Include="ManageStudentList.cs" />
     <Compile Include="ManageTeacher.cs" />
-    <Compile Include="ManageUserThings.cs" />
+    <Compile Include="ManageStudent.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Student.cs" />
